fix: move translateFromNetwork helper methode from client to server side
