+++ src/main/java/appeng/client/gui/implementations/GuiInterfaceTerminal.java
@@ -24,6 +24,7 @@ import java.util.TreeMap;
 import java.util.TreeSet;
 import java.util.function.Predicate;
 
+import appeng.me.cluster.implementations.CraftingCPUCluster;
 import net.minecraft.client.gui.FontRenderer;
 import net.minecraft.client.gui.GuiButton;
 import net.minecraft.client.renderer.RenderHelper;
@@ -1331,7 +1332,7 @@ public class GuiInterfaceTerminal extends AEBaseGui
 
         InterfaceTerminalEntry(long id, String name, int rows, int rowSize, boolean online, boolean p2pOutput) {
             this.id = id;
-            this.dispName = translateFromNetwork(name);
+            this.dispName = CraftingCPUCluster.translateFromNetwork(name);
             this.inv = new AppEngInternalInventory(null, rows * rowSize, 1);
             this.rows = rows;
             this.rowSize = rowSize;
@@ -1520,18 +1521,4 @@ public class GuiInterfaceTerminal extends AEBaseGui
         }
     }
 
-    public static String translateFromNetwork(String name) {
-        final String dispName;
-        if (StatCollector.canTranslate(name)) {
-            dispName = StatCollector.translateToLocal(name);
-        } else {
-            String fallback = name + ".name"; // its whatever. save some bytes on network but looks ugly
-            if (StatCollector.canTranslate(fallback)) {
-                dispName = StatCollector.translateToLocal(fallback);
-            } else {
-                dispName = StatCollector.translateToFallback(name);
-            }
-        }
-        return dispName;
-    }
 }
