+++ src/main/java/appeng/me/cluster/implementations/CraftingCPUCluster.java
@@ -50,6 +50,7 @@ import net.minecraft.util.ChatComponentText;
 import net.minecraft.util.ChatComponentTranslation;
 import net.minecraft.util.EnumChatFormatting;
 import net.minecraft.util.IChatComponent;
+import net.minecraft.util.StatCollector;
 import net.minecraft.world.World;
 import net.minecraft.world.WorldServer;
 
@@ -98,7 +99,6 @@ import appeng.api.util.DimensionalCoord;
 import appeng.api.util.IInterfaceViewable;
 import appeng.api.util.NamedDimensionalCoord;
 import appeng.api.util.WorldCoord;
-import appeng.client.gui.implementations.GuiInterfaceTerminal;
 import appeng.container.ContainerNull;
 import appeng.container.implementations.ContainerCraftingCPU;
 import appeng.core.AELog;
@@ -1346,7 +1346,7 @@ public final class CraftingCPUCluster implements IAECluster, ICraftingCPU {
                                 for (ICraftingMedium craftingProvider : craftingProviders) {
                                     final TileEntity tile = this.getTile(craftingProvider);
                                     if (tile instanceof TileInterface tileInterface) {
-                                        final String dispName = GuiInterfaceTerminal.translateFromNetwork(
+                                        final String dispName = translateFromNetwork(
                                                 tileInterface.getInterfaceDuality().getTermName());
                                         dimensionalCoords.add(new NamedDimensionalCoord(tile, dispName));
                                     }
@@ -1774,6 +1774,21 @@ public final class CraftingCPUCluster implements IAECluster, ICraftingCPU {
         } else {
             countToTryExtractItems++;
         }
+    }
+
+    public static String translateFromNetwork(String name) {
+        final String dispName;
+        if (StatCollector.canTranslate(name)) {
+            dispName = StatCollector.translateToLocal(name);
+        } else {
+            String fallback = name + ".name"; // its whatever. save some bytes on network but looks ugly
+            if (StatCollector.canTranslate(fallback)) {
+                dispName = StatCollector.translateToLocal(fallback);
+            } else {
+                dispName = StatCollector.translateToFallback(name);
+            }
+        }
+        return dispName;
     }
 
     public BaseActionSource getCurrentJobSource() {
